renamed rxjs-example to fetchexample
