--- starters/cra-rxjs-styled-components/src/components/Fetch-Example/index.ts
@@ -1,1 +1,0 @@
-export { FetchExample } from './Fetch-Example';
