+++ starters/cra-rxjs-styled-components/src/components/FetchExample/index.ts
@@ -1,0 +1,1 @@
+export { FetchExample } from './FetchExample';
